refactor: 변경된 클래스명 컨벤션에 맞게 클래스명 변경
