+++ src/main/java/com/programmers/bucketback/domains/review/application/ReviewService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 
 import com.programmers.bucketback.domains.common.MemberUtils;
 import com.programmers.bucketback.domains.common.vo.CursorPageParameters;
-import com.programmers.bucketback.domains.review.application.dto.GetReviewByCursorServiceResponse;
-import com.programmers.bucketback.domains.review.application.dto.ReviewContent;
+import com.programmers.bucketback.domains.review.application.dto.ReviewGetByCursorServiceResponse;
+import com.programmers.bucketback.domains.review.application.vo.ReviewContent;
 
 import lombok.RequiredArgsConstructor;
 
@@ -38,7 +38,7 @@ public class ReviewService {
 		reviewModifier.modify(reviewId, reviewContent);
 	}
 
-	public GetReviewByCursorServiceResponse getReviewsByCursor(
+	public ReviewGetByCursorServiceResponse getReviewsByCursor(
 		final Long itemId,
 		final CursorPageParameters parameters
 	) {
