+++ src/main/java/com/programmers/bucketback/domains/review/application/dto/ReviewGetByCursorServiceResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.programmers.bucketback.domains.review.api.dto.response.ReviewGetByCursorResponse;
 import com.programmers.bucketback.domains.review.application.vo.ReviewCursorSummary;
 
-public record GetReviewByCursorServiceResponse(
+public record ReviewGetByCursorServiceResponse(
 	Long reviewCount,
 	String nextCursorId,
 	List<ReviewCursorSummary> reviews
